requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.9
django-bootstrap5==24.2
django-celery-results==2.5.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-registration==0.1.3
drf-spectacular==0.28.0
drf-yasg==1.21.7
exceptiongroup==1.2.2
flower==2.0.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
Markdown==3.7
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
pycparser==2.22
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
rabbitmq==0.2.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.0
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13